+++ src/Library/ICommandNotificationOptions.cs
@@ -1,0 +1,7 @@
+namespace Scheduler;
+
+public interface ICommandNotificationOptions
+{
+    string? CommandSucceeded { get; set; }
+    string? CommandFailed { get; set; }
+}